fixed mismatched id type used for repo and controller, added resourcenotfound custom exception, adjusted get and post calls in usercontroller

+++ src/main/java/com/cognixia/jump/exception/InvalidInputException.java
@@ -4,7 +4,7 @@ public class InvalidInputException extends Exception{
 	
 	private static final long serialVersionUID = 1L;
 
-	public InvalidInputException(int user_id) {
+	public InvalidInputException(long user_id) {
 		super("The user_id that you inputed is not on the server");
 	}
 }
